Added hare main
