+++ .gitignore
@@ -1,6 +1,7 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
 /tmp
+/app/tmp
 */**/*un~
 */**/*.test
 */**/temp.js
