+++ app/cmd/root.go
@@ -1,9 +1,11 @@
 package cmd
 
 import (
+	"fmt"
 	cfg "github.com/spacemeshos/go-spacemesh/config"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"os"
 )
 
 var (
@@ -17,100 +19,110 @@ var RootCmd = &cobra.Command{
 }
 
 func init() {
+	addCommands(RootCmd)
+}
+
+func addCommands(cmd *cobra.Command) {
 
 	/** ======================== BaseConfig Flags ========================== **/
-	RootCmd.PersistentFlags().StringVarP(&config.BaseConfig.ConfigFile,
+	cmd.PersistentFlags().StringVarP(&config.BaseConfig.ConfigFile,
 		"config", "c", config.BaseConfig.ConfigFile, "Set Load configuration from file")
-	RootCmd.PersistentFlags().StringVarP(&config.BaseConfig.DataDir, "data-folder", "d",
+	cmd.PersistentFlags().StringVarP(&config.BaseConfig.DataDir, "data-folder", "d",
 		config.BaseConfig.DataDir, "Specify data directory for spacemesh")
-	RootCmd.PersistentFlags().BoolVar(&config.TestMode, "test-mode",
+	cmd.PersistentFlags().BoolVar(&config.TestMode, "test-mode",
 		config.TestMode, "Initialize testing features")
-	RootCmd.PersistentFlags().BoolVar(&config.CollectMetrics, "metrics",
+	cmd.PersistentFlags().BoolVar(&config.CollectMetrics, "metrics",
 		config.CollectMetrics, "collect node metrics")
-	RootCmd.PersistentFlags().IntVar(&config.MetricsPort, "metrics-port",
+	cmd.PersistentFlags().IntVar(&config.MetricsPort, "metrics-port",
 		config.MetricsPort, "metric server port")
-	RootCmd.PersistentFlags().StringVar(&config.OracleServer, "oracle_server",
+	cmd.PersistentFlags().StringVar(&config.OracleServer, "oracle_server",
 		config.OracleServer, "The oracle server url. (temporary) ")
-	RootCmd.PersistentFlags().Uint64Var(&config.OracleServerWorldId, "oracle_server_worldid",
+	cmd.PersistentFlags().IntVar(&config.OracleServerWorldId, "oracle_server_worldid",
 		config.OracleServerWorldId, "The worldid to use with the oracle server (temporary) ")
-	RootCmd.PersistentFlags().StringVar(&config.GenesisTime, "genesis-time",
+	cmd.PersistentFlags().StringVar(&config.GenesisTime, "genesis-time",
 		config.GenesisTime, "Time of the genesis layer in 2019-13-02T17:02:00+00:00 format")
-	RootCmd.PersistentFlags().IntVar(&config.LayerDurationSec, "layer-duration-sec",
+	cmd.PersistentFlags().IntVar(&config.LayerDurationSec, "layer-duration-sec",
 		config.LayerDurationSec, "Duration between layers in seconds")
 	/** ======================== P2P Flags ========================== **/
-	RootCmd.PersistentFlags().IntVar(&config.P2P.SecurityParam, "security-param",
+	cmd.PersistentFlags().IntVar(&config.P2P.SecurityParam, "security-param",
 		config.P2P.SecurityParam, "Consensus protocol k security param")
-	RootCmd.PersistentFlags().IntVar(&config.P2P.TCPPort, "tcp-port",
+	cmd.PersistentFlags().IntVar(&config.P2P.TCPPort, "tcp-port",
 		config.P2P.TCPPort, "TCP Port to listen on")
-	RootCmd.PersistentFlags().DurationVar(&config.P2P.DialTimeout, "dial-timeout",
+	cmd.PersistentFlags().DurationVar(&config.P2P.DialTimeout, "dial-timeout",
 		config.P2P.DialTimeout, "Network dial timeout duration")
-	RootCmd.PersistentFlags().DurationVar(&config.P2P.ConnKeepAlive, "conn-keepalive",
+	cmd.PersistentFlags().DurationVar(&config.P2P.ConnKeepAlive, "conn-keepalive",
 		config.P2P.ConnKeepAlive, "Network connection keep alive")
-	RootCmd.PersistentFlags().Int8Var(&config.P2P.NetworkID, "network-id",
+	cmd.PersistentFlags().Int8Var(&config.P2P.NetworkID, "network-id",
 		config.P2P.NetworkID, "NetworkID to run on (0 - mainnet, 1 - testnet)")
-	RootCmd.PersistentFlags().DurationVar(&config.P2P.ResponseTimeout, "response-timeout",
+	cmd.PersistentFlags().DurationVar(&config.P2P.ResponseTimeout, "response-timeout",
 		config.P2P.ResponseTimeout, "Timeout for waiting on resposne message")
-	RootCmd.PersistentFlags().StringVar(&config.P2P.NodeID, "node-id",
+	cmd.PersistentFlags().StringVar(&config.P2P.NodeID, "node-id",
 		config.P2P.NodeID, "Load node data by id (pub key) from local store")
-	RootCmd.PersistentFlags().BoolVar(&config.P2P.NewNode, "new-node",
+	cmd.PersistentFlags().BoolVar(&config.P2P.NewNode, "new-node",
 		config.P2P.NewNode, "Load node data by id (pub key) from local store")
-	RootCmd.PersistentFlags().IntVar(&config.P2P.BufferSize, "buffer-size",
+	cmd.PersistentFlags().IntVar(&config.P2P.BufferSize, "buffer-size",
 		config.P2P.BufferSize, "Size of the messages handler's buffer")
-	RootCmd.PersistentFlags().BoolVar(&config.P2P.SwarmConfig.Gossip, "gossip",
+	cmd.PersistentFlags().BoolVar(&config.P2P.SwarmConfig.Gossip, "gossip",
 		config.P2P.SwarmConfig.Gossip, "should we start a gossiping node?")
-	RootCmd.PersistentFlags().BoolVar(&config.P2P.SwarmConfig.Bootstrap, "bootstrap",
+	cmd.PersistentFlags().BoolVar(&config.P2P.SwarmConfig.Bootstrap, "bootstrap",
 		config.P2P.SwarmConfig.Bootstrap, "Bootstrap the swarm")
-	RootCmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RoutingTableBucketSize, "bucketsize",
+	cmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RoutingTableBucketSize, "bucketsize",
 		config.P2P.SwarmConfig.RoutingTableBucketSize, "The rounding table bucket size")
-	RootCmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RoutingTableAlpha, "alpha",
+	cmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RoutingTableAlpha, "alpha",
 		config.P2P.SwarmConfig.RoutingTableAlpha, "The rounding table Alpha")
-	RootCmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RandomConnections, "randcon",
+	cmd.PersistentFlags().IntVar(&config.P2P.SwarmConfig.RandomConnections, "randcon",
 		config.P2P.SwarmConfig.RoutingTableAlpha, "Number of random connections")
-	RootCmd.PersistentFlags().StringSliceVar(&config.P2P.SwarmConfig.BootstrapNodes, "bootnodes",
+	cmd.PersistentFlags().StringSliceVar(&config.P2P.SwarmConfig.BootstrapNodes, "bootnodes",
 		config.P2P.SwarmConfig.BootstrapNodes, "Number of random connections")
-	RootCmd.PersistentFlags().DurationVar(&config.TIME.MaxAllowedDrift, "max-allowed-time-drift",
+	cmd.PersistentFlags().DurationVar(&config.TIME.MaxAllowedDrift, "max-allowed-time-drift",
 		config.TIME.MaxAllowedDrift, "When to close the app until user resolves time sync problems")
-	RootCmd.PersistentFlags().IntVar(&config.TIME.NtpQueries, "ntp-queries",
+	cmd.PersistentFlags().IntVar(&config.TIME.NtpQueries, "ntp-queries",
 		config.TIME.NtpQueries, "Number of ntp queries to do")
-	RootCmd.PersistentFlags().DurationVar(&config.TIME.DefaultTimeoutLatency, "default-timeout-latency",
+	cmd.PersistentFlags().DurationVar(&config.TIME.DefaultTimeoutLatency, "default-timeout-latency",
 		config.TIME.DefaultTimeoutLatency, "Default timeout to ntp query")
-	RootCmd.PersistentFlags().DurationVar(&config.TIME.RefreshNtpInterval, "refresh-ntp-interval",
+	cmd.PersistentFlags().DurationVar(&config.TIME.RefreshNtpInterval, "refresh-ntp-interval",
 		config.TIME.RefreshNtpInterval, "Refresh intervals to ntp")
 
 	/** ======================== API Flags ========================== **/
 	// StartJSONApiServerFlag determines if json api server should be started
-	RootCmd.PersistentFlags().BoolVar(&config.API.StartJSONServer, "json-server",
+	cmd.PersistentFlags().BoolVar(&config.API.StartJSONServer, "json-server",
 		config.API.StartJSONServer, "StartService the json http server. "+
 			"Note that starting the Json server also starts the grpc server.",
 	)
 	// JSONServerPortFlag determines the json api server local listening port
-	RootCmd.PersistentFlags().IntVar(&config.API.JSONServerPort, "json-port",
+	cmd.PersistentFlags().IntVar(&config.API.JSONServerPort, "json-port",
 		config.API.JSONServerPort, "JSON api server port")
 	// StartGrpcAPIServerFlag determines if the grpc server should be started
-	RootCmd.PersistentFlags().BoolVar(&config.API.StartGrpcServer, "grpc-server",
+	cmd.PersistentFlags().BoolVar(&config.API.StartGrpcServer, "grpc-server",
 		config.API.StartGrpcServer, "StartService the grpc server")
 	// GrpcServerPortFlag determines the grpc server local listening port
-	RootCmd.PersistentFlags().IntVar(&config.API.GrpcServerPort, "grpc-port",
+	cmd.PersistentFlags().IntVar(&config.API.GrpcServerPort, "grpc-port",
 		config.API.GrpcServerPort, "GRPC api server port")
 
 	/**========================Hare Flags ========================== **/
 
 	// N determines the size of the hare committee
-	RootCmd.PersistentFlags().IntVar(&config.HARE.N, "hare-committee-size",
+	cmd.PersistentFlags().IntVar(&config.HARE.N, "hare-committee-size",
 		config.HARE.N, "Size of Hare committee")
 	// F determines the max number of adversaries in the Hare committee
-	RootCmd.PersistentFlags().IntVar(&config.HARE.F, "hare-max-adversaries",
+	cmd.PersistentFlags().IntVar(&config.HARE.F, "hare-max-adversaries",
 		config.HARE.F, "Max number of adversaries in the Hare committee")
 	// RoundDuration determines the duration of a round in the Hare protocol
-	RootCmd.PersistentFlags().DurationVar(&config.HARE.RoundDuration, "hare-round-duration-ms",
+	cmd.PersistentFlags().DurationVar(&config.HARE.RoundDuration, "hare-round-duration-ms",
 		config.HARE.RoundDuration, "Duration of round in the Hare protocol")
 
 	/**========================Consensus Flags ========================== **/
-	//todo: add this here
-
-	RootCmd.AddCommand(VersionCmd)
 
 	// Bind Flags to config
-	viper.BindPFlags(RootCmd.PersistentFlags())
+	viper.BindPFlags(cmd.PersistentFlags())
 
+}
+
+// Execute adds all child commands to the root command sets flags appropriately.
+// This is called by main.main(). It only needs to happen once to the rootCmd.
+func Execute() {
+	if err := RootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(-1)
+	}
 }
